% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Kibior equals operator}
\alias{Kibior equals operator}
\alias{==.KibiorOperators}
\title{Kibior equals operator}
\usage{
\method{==}{KibiorOperators}(x, y)
}
\arguments{
\item{x}{the first Kibior instance}

\item{y}{the second Kibior instance}
}
\value{
TRUE if the two instances are equals, else FALSE
}
\description{
Kibior equals operator
}
\details{
Call kibioR `$eq()` for a comparison of the two instances.
}
\seealso{
Other comparison: 
\code{\link{Kibior not-equals operator}},
\code{\link{Static - Kibior is instance}}
}
\concept{comparison}
