% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Kibior not-equals operator}
\alias{Kibior not-equals operator}
\alias{!=.KibiorOperators}
\title{Kibior not-equals operator}
\usage{
\method{!=}{KibiorOperators}(x, y)
}
\arguments{
\item{x}{the first Kibior instance}

\item{y}{the second Kibior instance}
}
\value{
TRUE if the two instances are differents, else FALSE
}
\description{
Kibior not-equals operator
}
\details{
Call kibioR `$ne()` for a comparison of the two instances.
}
\seealso{
Other comparison: 
\code{\link{Kibior equals operator}},
\code{\link{Static - Kibior is instance}}
}
\concept{comparison}
