% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Static - initiate a direct instance to Kibio public repository}
\alias{Static - initiate a direct instance to Kibio public repository}
\title{Static - initiate a direct instance to Kibio public repository}
\arguments{
\item{verbose}{verbosity activation (default: FALSE)}
}
\value{
a new instance of Kibior conencted to Kibio service
}
\description{
Static - initiate a direct instance to Kibio public repository
}
\details{
Initiate a instance of Kibior connected to the Kibio public repository.
}
\concept{initiate}
