% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Static - Install packages}
\alias{Static - Install packages}
\title{Static - Install packages}
\arguments{
\item{pkg_names}{a vector of some package names}

\item{install}{a logical thtat defines if packages should be installed. (default: TRUE)}
}
\description{
Static - Install packages
}
\details{
Check packages installation in the current env from a given list
}
\seealso{
Other install: 
\code{\link{Static - Tests if packages are installed}}
}
\concept{install}
