% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Static - Kibior is instance}
\alias{Static - Kibior is instance}
\title{Static - Kibior is instance}
\arguments{
\item{obj}{an object}
}
\value{
TRUE if the given object is an instance of Kibior, else FALSE
}
\description{
Static - Kibior is instance
}
\details{
Tests if a given object is a Kibior instance.
Basically compute symmetric difference between two sets of class.
}
\seealso{
Other comparison: 
\code{\link{Kibior equals operator}},
\code{\link{Kibior not-equals operator}}
}
\concept{comparison}
