% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Static - Tests if packages are installed}
\alias{Static - Tests if packages are installed}
\title{Static - Tests if packages are installed}
\arguments{
\item{pkg_names}{a vector of some package names}
}
\value{
a named vector of packages with installation status
}
\description{
Static - Tests if packages are installed
}
\details{
Get the installation status of some package names (installed TRUE/FALSE).
}
\concept{install}
