% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{[static] Install packages}
\alias{[static] Install packages}
\title{[static] Install packages}
\arguments{
\item{pkg_names}{a vector of some package names}

\item{install}{a logical thtat defines if packages should be installed. (default: TRUE)}
}
\description{
[static] Install packages
}
\details{
Check packages installation in the current env from a given list
}
\seealso{
Other install: 
\code{\link{[static] Tests if packages are installed}}
}
\concept{install}
