% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{[static] Tests if packages are installed}
\alias{[static] Tests if packages are installed}
\title{[static] Tests if packages are installed}
\arguments{
\item{pkg_names}{a vector of some package names}
}
\value{
a vector of not installed packages
}
\description{
[static] Tests if packages are installed
}
\details{
Given a vector of string, returns a list of packages that are not installed 
in the current env
}
\seealso{
Other install: 
\code{\link{[static] Install packages}}
}
\concept{install}
