% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Static - Tests if packages are installed}
\alias{Static - Tests if packages are installed}
\title{Static - Tests if packages are installed}
\arguments{
\item{pkg_names}{a vector of some package names}
}
\value{
a vector of not installed packages
}
\description{
Static - Tests if packages are installed
}
\details{
Given a vector of string, returns a list of packages that are not installed 
in the current env
}
\seealso{
Other install: 
\code{\link{Static - Install packages}}
}
\concept{install}
